Refact: Modify component file structure
