+++ src/app/my-page/message/panel/MessageList.tsx
@@ -1,91 +1,7 @@
 'use client'
-import { useRouter } from 'next/navigation'
-import dayjs from 'dayjs'
-import {
-  Avatar,
-  Checkbox,
-  Chip,
-  List,
-  ListItem,
-  ListItemAvatar,
-  ListItemButton,
-  ListItemIcon,
-  ListItemSecondaryAction,
-  ListItemText,
-  Stack,
-  Typography,
-} from '@mui/material'
+import { List, Typography } from '@mui/material'
 import { IMessageListData } from '@/types/IMessage'
-
-interface IMessageItemProps {
-  message: IMessageListData
-  isManageMode: boolean
-  isChecked: boolean
-  toggleSelectUser: (targetId: number) => void
-}
-
-const PCMessageItem = ({
-  message,
-  isManageMode,
-  isChecked,
-  toggleSelectUser,
-}: IMessageItemProps) => {
-  const router = useRouter()
-  const label = { inputProps: { 'aria-label': 'MessageItem Checkbox' } }
-  const { targetId, conversationId } = message
-  return (
-    <ListItem>
-      <ListItemButton
-        role={undefined}
-        onClick={
-          isManageMode
-            ? () => toggleSelectUser(targetId)
-            : () =>
-                router.push(
-                  `/my-page/message/${conversationId}?target=${targetId}`,
-                )
-        }
-      >
-        {isManageMode && (
-          <ListItemIcon>
-            <Checkbox {...label} checked={isChecked} />
-          </ListItemIcon>
-        )}
-        <MessageItemBase message={message} />
-      </ListItemButton>
-    </ListItem>
-  )
-}
-
-const MessageItemBase = ({ message }: { message: IMessageListData }) => {
-  const { targetNickname, latestContent, targetProfile, unreadMsgNumber } =
-    message
-  return (
-    <>
-      <ListItemAvatar>
-        <Avatar src={targetProfile} />
-      </ListItemAvatar>
-      <ListItemText primary={targetNickname} secondary={latestContent} />
-      <ListItemSecondaryAction>
-        <Stack alignItems={'flex-end'}>
-          <Typography>
-            {/* TODO : latestDate 데이터로 바꿀 것!!!!!!! */}
-            {dayjs('2021-08-01T16:26:39.098').format('MM월 DD일')}
-          </Typography>
-          <Chip
-            label={
-              unreadMsgNumber > 0
-                ? unreadMsgNumber > 99
-                  ? '99+'
-                  : unreadMsgNumber
-                : null
-            }
-          />
-        </Stack>
-      </ListItemSecondaryAction>
-    </>
-  )
-}
+import { PCMessageItem } from './MessageItem'
 
 interface IMessageListProps {
   messageList: IMessageListData[]
